2013-09-25 Miguel de Dios <miguel.dedios@artica.es>
	* godmode/modules/manage_network_components_form_common.php,
	godmode/modules/manage_network_components_form.php: fixed the
	javascript to configure the form of snmp.
	
	* godmode/agentes/module_manager_editor.php,
	godmode/agentes/configurar_agente.php,
	godmode/agentes/module_manager_editor_common.php: fixed the
	"module_macros" when it is fail when is not set for example.




git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@8806 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/godmode/agentes/module_manager_editor_common.php
@@ -483,13 +483,17 @@ $table_macros->style[5] = 'width: 10px';
 $table_macros->colspan = array ();
 
 $macro_count = 0;
-foreach ($module_macros as $macro_name => $macro_value) {
-	$table_macros->data[$macro_count][0] = __('Name');
-	$table_macros->data[$macro_count][1] = html_print_input_text ('module_macro_names[]', $macro_name, '', 50, 60, true);
-	$table_macros->data[$macro_count][2] = __('Value');
-	$table_macros->data[$macro_count][3] = html_print_input_text ('module_macro_values[]', $macro_value, '', 50, 60, true);
-	$table_macros->data[$macro_count][4] = '<a href="javascript: delete_macro(' . $macro_count . ');">' . html_print_image('images/cross.png', true) . '</a>';
-	$macro_count++;
+if (isset($module_macros)) {
+	if (is_array($module_macros)) {
+		foreach ($module_macros as $macro_name => $macro_value) {
+			$table_macros->data[$macro_count][0] = __('Name');
+			$table_macros->data[$macro_count][1] = html_print_input_text ('module_macro_names[]', $macro_name, '', 50, 60, true);
+			$table_macros->data[$macro_count][2] = __('Value');
+			$table_macros->data[$macro_count][3] = html_print_input_text ('module_macro_values[]', $macro_value, '', 50, 60, true);
+			$table_macros->data[$macro_count][4] = '<a href="javascript: delete_macro(' . $macro_count . ');">' . html_print_image('images/cross.png', true) . '</a>';
+			$macro_count++;
+		}
+	}
 }
 $table_macros->data[$macro_count][0] = '<span>'.__('Add module macro').'</span> <a href="javascript:add_macro();">'.html_print_image('images/add.png',true).'</a>';
 $table_macros->colspan[$macro_count][0] = 5;
